+++ addons/snailmail/wizard/__init__.py
@@ -1,3 +1,2 @@
-from . import snailmail_confirm
 from . import snailmail_letter_format_error
 from . import snailmail_letter_missing_required_fields
