+++ addons/snailmail_account/wizard/__init__.py
@@ -1,4 +1,4 @@
 # -*- coding: utf-8 -*-
 
 from . import account_invoice_send
-from . import snailmail_confirm_invoice_send
+from . import snailmail_confirm_invoice
