+++ addons/snailmail_account/wizard/snailmail_confirm_invoice_views.xml
@@ -1,11 +1,12 @@
 <?xml version="1.0" encoding="utf-8"?>
 <odoo>
-    <record id="snailmail_confirm_view" model="ir.ui.view">
-        <field name="name">snailmail.confirm.view</field>
-        <field name="model">snailmail.confirm</field>
+    <record id="snailmail_confirm_invoice_view_form" model="ir.ui.view">
+        <field name="name">snailmail.confirm.invoice.view.form</field>
+        <field name="model">snailmail.confirm.invoice</field>
         <field name="arch" type="xml">
             <form string="Snailmail Confirmation">
-                <p>You are about to send this <field name="model_name" readonly="1" class="oe_inline" /> by post. Are you sure you want to continue?</p>
+                <field name="invoice_send_id" invisible="1"/>
+                <p>You are about to send this Invoice by post. Are you sure you want to continue?</p>
                 <footer>  
                     <button string="Confirm" name="action_confirm" type="object" class="btn-primary" data-hotkey="q"/>
                     <button string="Cancel" name="action_cancel" type="object" class="btn-secondary" data-hotkey="x"/>
